[Frontend] - Переименованны маршруты, добаленно добаление очков для вариантов ответа и формирование максимальной оценки за прохождение опроса. [Backend] - Добавленная логика авторизации пользователя, обработки ответов пользователя, формирование статистики.

+++ backend/server/quizapi/views.py
@@ -2,27 +2,20 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.exceptions import AuthenticationFailed
 from rest_framework import status
-from .models import Quiz, User, Question, QuestionGroup
+from .models import Quiz, User, Question, QuestionGroup, UserAnswers, QuizStatistic
 from . import serializers
-import datetime, jwt
-
-# TODO add jwt token checking for all routes
-# def check_manager_access(cookies = None):
-#     if cookies is None:
-#         return False
-#
-#     token = cookies.get('jwt')
-#     payload = jwt.decode(token, 'secret', algorithms=['HS256'])
-#
-#     if payload['status'] != 'manager':
-#         return False
-#     return True
+import json, jwt
+from .jwt_methods import get_data_from_jwt, set_jwt
 
 class QuizView(APIView):
     def post(self, request):
         """
         Создание опроса
         """
+        # payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        # if payload is False or payload['status'] != 'manager':
+        #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
         serialized_quiz = serializers.QuizSerializer(data=request.data)
         if serialized_quiz.is_valid():
             serialized_quiz.save()
@@ -38,61 +31,69 @@ class QuizView(APIView):
     def get(self, request):
         """
         Получение списка всех опросов
+        :return: Объект из двух списокв:
+                quiz_list - список не пройденных опросов,
+                done_quiz_list - список пройденных вопросов
         """
-        quiz_list = Quiz.objects.all()
+
+        payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        # if payload is False or payload['status'] != 'worker':
+        #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
+        user_answers = UserAnswers.objects.filter(user_id=payload['id'])
+        serialized_answers = serializers.UserAnswersSerializer(user_answers, many=True)
+        user_answers = json.loads(json.dumps(serialized_answers.data))
+
+        quiz_ids = []
+        for answer in user_answers:
+            quiz_ids.append(answer['quiz_id'])
+
+        quiz_list = Quiz.objects.exclude(id__in=quiz_ids)
         serialized_quiz = serializers.QuizSerializer(quiz_list, many=True)
-        serialized_quiz_copy = list(serialized_quiz.data).copy()
+        quiz_list = json.loads(json.dumps(serialized_quiz.data))
 
-        # TODO try to change this algorithm
-        for quiz in serialized_quiz_copy:
-            for key in quiz:
-                if key == 'questions':
-                    for question in quiz['questions']:
-                        for answer in list(question):
-                            if answer == 'answer':
-                                del question[answer]
+        if len(quiz_list) != 0:
+            for quiz in quiz_list:
+                del quiz['quiz_max_grade']
+                for question in quiz['questions']:
+                    del question['answer']
+                    del question['question_max_grade']
 
-        return Response(serialized_quiz_copy)
+                    for variant in question['variants']:
+                        del variant['score']
+
+        return Response(status=status.HTTP_200_OK, data={
+            "quiz_list": quiz_list,
+            "done_quiz_list": user_answers
+        })
 
 class QuestionListView(APIView):
-    # TODO this logic can be done only if request was made by manager also create api for making questions groups
 
     def get(self, request):
-        # token = request.COOKIES.get('jwt')
-        # payload = jwt.decode(token, 'secret', algorithms=['HS256'])
-        #
-        # if payload['status'] != 'manager':
+        # payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        # if payload is False or payload['status'] != 'manager':
         #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
 
         questions_list = Question.objects.all()
         serialized_questions = serializers.QuestionSerializer(questions_list, many=True)
         return Response(status=status.HTTP_200_OK, data=serialized_questions.data)
 
-class CheckingUserAnswersView(APIView):
-    """
-    Оцека ответов на опрос
-    """
-    def post(self, request):
-        quiz_id = request.data.pop('quizId')
-        answers = request.data.pop('answers')
-
-        for answer in answers:
-            question_id = answer.pop('questionId')
-            question = Question.objects.all().filter(id=question_id).first()
-            serialized_question = serializers.QuestionSerializer(question)
-            # TODO поиск подстроки ответа пользователя в строке ответа на вопрос
-            # TODO формирование словаря
-            # { "id_вопроса": id, "question": вопрос, "user_answer": ответ_пользователя, "user_grade": баллы_за_вопрос, "max_grade": максимальный_балл }
-            pass
-
-        return Response()
-
 class RegisterView(APIView):
     def post(self, request):
+        """
+        Регистрайия пользователя
+        """
         new_user = serializers.UserSerializer(data=request.data)
         new_user.is_valid(raise_exception=True)
         new_user.save()
-        return Response(status=status.HTTP_201_CREATED, data=new_user.data)
+
+        token = set_jwt(new_user.data['id'], request.data['auth_status'])
+
+        response = Response()
+        response.set_cookie(key='jwt', value=token, httponly=True)
+        response.data = new_user.data
+
+        return response
 
 class UserView(APIView):
     def get(self, request):
@@ -102,7 +103,7 @@ class UserView(APIView):
             raise AuthenticationFailed('Unauthenticated')
 
         try:
-           payload = jwt.decode(token, 'secret', algorithms=['HS256'])
+           payload = get_data_from_jwt(token)
         except jwt.ExpiredSignatureError:
             raise AuthenticationFailed('Unauthenticated')
 
@@ -114,7 +115,7 @@ class LoginView(APIView):
     def post(self, request):
         email = request.data['email']
         password = request.data['password']
-        is_manager = request.data['is_manager']
+        auth_status = request.data['auth_status']
 
         user = User.objects.filter(email=email).first()
 
@@ -124,22 +125,14 @@ class LoginView(APIView):
         if not user.check_password(password):
             raise AuthenticationFailed('Incorrect password')
 
-        if user.is_manager != is_manager:
+        if user.auth_status != auth_status:
             raise AuthenticationFailed('Incorrect authentication status')
 
-        payload = {
-            'id': user.id,
-            'exp': datetime.datetime.utcnow() + datetime.timedelta(minutes=60),
-            'iat': datetime.datetime.utcnow(),
-            'status': user.is_manager
-        }
-
-        token = jwt.encode(payload, 'secret', algorithm='HS256')
+        token = set_jwt(user.id, user.auth_status)
 
         response = Response()
         response.set_cookie(key='jwt', value=token, httponly=True)
         response.data = {
-            'jwt': token,
             "name": user.name,
             "surname": user.surname,
             "email": user.email
@@ -158,6 +151,10 @@ class LogoutView(APIView):
 # Question Group
 class QuestionGroupView(APIView):
     def get(self, request):
+        # payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        # if payload is False or payload['auth_status'] != 'manager':
+        #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
         groups = QuestionGroup.objects.all()
         serialized_groups = serializers.QuestionGroupSerializer(groups, many=True)
         return Response(data=serialized_groups.data)
@@ -166,6 +163,10 @@ class QuestionGroupView(APIView):
         """
         Создание новой группы вопросов
         """
+        # payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        # if payload is False or payload['auth_status'] != 'manager':
+        #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
         new_group = serializers.QuestionGroupSerializer(data=request.data)
         new_group.is_valid(raise_exception=True)
         new_group.save()
@@ -175,6 +176,10 @@ class QuestionGroupView(APIView):
         """
         Добавление вопросов в группу
         """
+        # payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        # if payload is False or payload['auth_status'] != 'manager':
+        #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
         group_name = request.data['group_name']
         group = QuestionGroup.objects.all().filter(group_name=group_name).first()
         updated_group = serializers.QuestionGroupSerializer(instance=group, data=request.data)
@@ -187,7 +192,105 @@ class QuestionGroupView(APIView):
         """
         Удаление группы вопросов
         """
+        payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        if payload is False or payload['auth_status'] != 'manager':
+            return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
         QuestionGroup.objects.filter(id=request.data.pop('group_id')).delete()
         groups = QuestionGroup.objects.all()
         serialized_group = serializers.QuestionGroupSerializer(groups, many=True)
         return Response(data=serialized_group.data)
+
+# Оценка прохождения опроса и формирование статистики
+class GradingUser(APIView):
+    def post(self, request):
+        quiz_id = request.data['quizId']
+        answers = request.data['answers']
+
+        payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+
+        # if payload is False or payload['status'] != 'worker':
+        #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
+        quiz_statistic = {}
+
+        quiz_instance = Quiz.objects.all().filter(id=quiz_id).first()
+        serialized_quiz = serializers.QuizSerializer(quiz_instance)
+        quiz_instance = json.loads(json.dumps(serialized_quiz.data))
+        quiz_statistic['quiz_name'] = quiz_instance['quiz_name']
+
+        user_instance = User.objects.filter(id=payload['id']).first()
+        serialized_user = serializers.UserSerializer(user_instance)
+        quiz_statistic['user_email'] = serialized_user.data['email']
+
+        quiz_name = quiz_instance['quiz_name']
+        quiz_max_grade = quiz_instance['quiz_max_grade']
+        quiz_statistic['quiz_max_grade'] = quiz_max_grade
+        user_score = 0
+
+        def searchEl(variant, variant_instance):
+            for item in variant_instance:
+                if item['variant'] == variant:
+                    return item['score']
+
+        question_statistic = []
+
+        for answer in answers:
+            question_id = answer.pop('questionId')
+            question_instance = Question.objects.all().filter(id=question_id).first()
+            serialized_question = serializers.QuestionSerializer(question_instance)
+            question_instance = json.loads(json.dumps(serialized_question.data))
+            question_user_grade = 0
+
+            for user_answer in answer['answer'].split(','):
+                exist = question_instance['answer'].find(user_answer)
+
+                if exist != -1:
+                    user_score += searchEl(user_answer, question_instance['variants'])
+                    question_user_grade += searchEl(user_answer, question_instance['variants'])
+
+            question_statistic.append({
+                "question_name": question_instance['question'],
+                "user_answer": answer['answer'],
+                "correct_answer": question_instance['answer'],
+                "user_grade": question_user_grade,
+                "question_max_grade": question_instance['question_max_grade'],
+            })
+
+            question_user_grade = 0
+
+        quiz_statistic['questions_statistic'] = question_statistic
+        quiz_statistic['user_grade'] = user_score
+
+        quiz_statistic_instance = serializers.QuizStatisticSerializer(data=quiz_statistic)
+        quiz_statistic_instance.is_valid(raise_exception=True)
+        quiz_statistic_instance.save()
+
+        user_answer = {
+            "user_id": payload['id'],
+            "quiz_id": quiz_id,
+            "quiz_name": quiz_name,
+            "user_grade": user_score,
+            "max_grade": quiz_max_grade
+        }
+
+        user_answer_instance = serializers.UserAnswersSerializer(data=user_answer)
+        user_answer_instance.is_valid(raise_exception=True)
+        user_answer_instance.save()
+
+        return Response(status=status.HTTP_201_CREATED, data={
+            "message": "Replies saved successfully",
+            "statistic": quiz_statistic
+        })
+
+
+# Отправка менеджеру полной статистики прохождения опросов
+class StatisticView(APIView):
+    def get(self, request):
+        # payload = get_data_from_jwt(request.COOKIES.get('jwt'))
+        # if payload is False or payload['auth_status'] != 'manager':
+        #     return Response(status=status.HTTP_401_UNAUTHORIZED, data={'message': 'Incorrect authorization'})
+
+        quiz_statistic_instance = QuizStatistic.objects.all()
+        serialized_statistic = serializers.QuizStatisticSerializer(quiz_statistic_instance, many=True)
+        return Response(status=status.HTTP_200_OK, data=serialized_statistic.data)
